Project 1: Final Submission
Attached is the preprocessing of the the Economic Freedom Dateset. This dateset has a multitude of different variables, identifying different aspects which are considered to be vital to identifying Economic Freedom. 

Additionally there is a mini experiment in order to identify the the exact effects of standardization and how the order in which they're used effects the subsequent model predictions. 

Abstract : 
Intituively splitting a dataset then standardizing both the trianing and test set with the same values seems to be some what abritrary as well as repetative. This is explained as to prevent bias from developing as if to guide a specific model into a specific outcome. Although it does cause question whether this statement of bias is true or not. The issue that I briefly cover is whether applying standardization towards a datasets before or after splitting effects the results of a model being run. I believe that this will effect the model for better. Using SKlearn's StandardScaler as the means of standardizing the datasets, the results showed that Standardizing the full datasets before the split produced a .976 F1 Score, Standardizing after the split using the Training data as the means to form said standardization created a worse result of 0.86 F1. While the control with no standardization was 0.974 as well. The model being used was KNN, although futhur experiments specifically identifying different ML models would have to be run to prove this brief result. These results make sense since the StandardScaler would use the X_train data and as a result this isn't fully representative of the data that it will be modeled, specifically the X_test set.